+++ image-blp/src/parser/direct/blp2.rs
@@ -17,9 +17,9 @@ pub fn parse_raw3<'a>(
     images: &mut Vec<Raw3Image>,
     input: &'a [u8],
 ) -> Parser<'a, ()> {
-    let mut read_image = |i: u32| {
-        let offset = offsets[i as usize];
-        let size = sizes[i as usize];
+    let mut read_image = |i: usize| {
+        let offset = offsets[i];
+        let size = sizes[i];
         if offset as usize >= original_input.len() {
             error!(
                 "Offset of mipmap {} is out of bounds! {} >= {}",
@@ -76,9 +76,9 @@ pub fn parse_dxt1<'a>(
 ) -> Parser<'a, ()> {
     trace!("{:?}", blp_header);
 
-    let mut read_image = |i: u32| {
-        let offset = offsets[i as usize];
-        let size = sizes[i as usize];
+    let mut read_image = |i: usize| {
+        let offset = offsets[i];
+        let size = sizes[i];
         if offset as usize >= original_input.len() {
             error!(
                 "Offset of mipmap {} is out of bounds! {} >= {}",
@@ -129,9 +129,9 @@ pub fn parse_dxt3<'a>(
 ) -> Parser<'a, ()> {
     trace!("{:?}", blp_header);
 
-    let mut read_image = |i: u32| {
-        let offset = offsets[i as usize];
-        let size = sizes[i as usize];
+    let mut read_image = |i: usize| {
+        let offset = offsets[i];
+        let size = sizes[i];
         if offset as usize >= original_input.len() {
             error!(
                 "Offset of mipmap {} is out of bounds! {} >= {}",
@@ -182,9 +182,9 @@ pub fn parse_dxt5<'a>(
 ) -> Parser<'a, ()> {
     trace!("{:?}", blp_header);
 
-    let mut read_image = |i: u32| {
-        let offset = offsets[i as usize];
-        let size = sizes[i as usize];
+    let mut read_image = |i: usize| {
+        let offset = offsets[i];
+        let size = sizes[i];
         if offset as usize >= original_input.len() {
             error!(
                 "Offset of mipmap {} is out of bounds! {} >= {}",
